data_evolution/pyproject.toml
[project]
name = "data-evolution"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9,<3.13"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.9.4",
    "pandas>=2.3.2",
    "pyyaml>=6.0.2",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
